plateforme_evaluation/requirement.txt
celery
djangorestframework
django-cors-headers
django-allauth
django-mfa
django-fsm
django-crispy-forms
redis
flower
channels
channels-redis
reportlab
pypdf
python-docx
crispy-bootstrap5==2023.10
crispy-bootstrap5
PyPDF2
ollama
django-cors-headers
PyJWT
django
djangorestframework
django-cors-headers
Pillow 
pytest
pytest-django
python-dotenv
psycopg2-binary
django-allauth
django-oauth-toolkit
pillow
python-pptx
django-mathfilters
pyserial
tqdm
cryptography
channels
redis